change memory release strategy
the memory release strategy used to be to give priority to the memory
queue
to pull the next priority. This behaviour made sense at one time, but no
longer makes any sense.

The new strategy is to just put things in the ready queue, and pull the
next process from there.